# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Trilix Atlassian Assistant
  description: Use Atlassian tools (Confluence, Jira) via Trilix MCP Server. Optimized for ChatGPT (30 ops).
  version: 1.0.0
servers:
  - url: https://candra-unsabred-totally.ngrok-free.dev
    description: Production Server
paths:
  /api/workspaces:
    get:
      operationId: listWorkspaces
      summary: List available connected workspaces
      description: Returns a list of workspaces the user has connected. You MUST provide your Clerk User ID (user_...) to retrieve your specific workspaces.
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: Your Clerk User ID for authentication (e.g., user_2pk...)
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceResponse"
    post:
      operationId: createWorkspace
      summary: Connect a new Atlassian workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
  /api/workspaces/{workspace_id}/status:
    get:
      operationId: getWorkspaceStatus
      summary: Check workspace connectivity status
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: Your Clerk User ID for authentication
      responses:
        "200":
          description: Workspace connection status
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaceId: { type: string }
                  connected: { type: boolean }
                  error: { type: string }

  # Confluence Tool Endpoints
...
components:
  schemas:
...
    CreateWorkspaceRequest:
      type: object
      properties:
        workspaceName: { type: string }
        siteUrl: { type: string }
        email: { type: string }
        apiToken: { type: string }
        user_id: { type: string, description: "Your Clerk User ID (required for impersonation)" }
      required: [siteUrl, email, apiToken, user_id]
    WorkspaceResponse:
      type: object
      properties:
        workspaceId: { type: string }
        workspaceName: { type: string }
        siteUrl: { type: string }
        email: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []